reorg test code
